More Poly1305 work
Fleshed out the initialization process. Started work on the block creation. Fixed an error in ChaCha where an unsigned short should have been an unsigned long long.

+++ src/include/chacha.h
@@ -222,7 +222,7 @@ void ChaChaEncryption::encryptAndDecryptSubProcess(char* message) {
 
 
 void ChaChaEncryption::encryptAndDecryptProcess(char* message, unsigned long long bytes, unsigned long startBlock = 0) {
-	for(unsigned short i = 0; i < (messageBlockCount - 1); i += 1) {
+	for(unsigned long long i = 0; i < (messageBlockCount - 1); i += 1) {
 		encryptAndDecryptSubProcess(message);
 		incrementBlockCounter();
 	}
